fix collaborators form again

+++ app/controllers/collaborations_controller.rb
@@ -3,19 +3,22 @@ class CollaborationsController < ApplicationController
 
   def create
     respond_to do |format|
-
       format.js do
+
         @wiki = Wiki.find(collaboration_params[:wiki_id])
-        @wiki.private = true
-        @wiki.save
         @collaborator = User.find_by(email: collaboration_params[:email])
 
         if @collaborator && @collaborator.collaborating_on?(@wiki)
           render :create, locals: { collaborator: @collaborator, wiki: @wiki, state: "exists" }
         elsif @collaborator
           @collaboration = Collaboration.new(wiki_id: collaboration_params[:wiki_id], user_id: @collaborator.id)
-          # authorize @collaboration
+          authorize @collaboration
           @collaboration.save
+
+          # Make sure wiki is private
+          @wiki.private = true
+          @wiki.save
+
           render :create, locals: { collaborator: @collaborator, wiki: @wiki, state: "new" }
         else
           render :create, locals: { collaborator: @collaborator, wiki: @wiki, state: "no_user"}
@@ -31,7 +34,7 @@ class CollaborationsController < ApplicationController
         @wiki = Wiki.find(collaboration_destroy_params[:wiki_id])
         @user = User.find(collaboration_destroy_params[:user_id])
         @collaboration = @wiki.collaborations.find_by(user_id: collaboration_destroy_params[:user_id])
-        # authorize @collaboration
+        authorize @collaboration
         
         if @collaboration.destroy
           render :destroy, locals: { user: @user}
